fix: some RoomData f32's should be u32's

+++ lib/botris/src/api.rs
@@ -117,8 +117,8 @@ pub struct RoomData {
     pub final_pps: f32,
     pub start_margin: f32,
     pub end_margin: f32,
-    pub ft: f32,
-    pub max_players: f32,
+    pub ft: u32,
+    pub max_players: u32,
     pub game_ongoing: bool,
     pub round_ongoing: bool,
     // startedAt: number | null;
